backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
faiss-cpu==1.8.0
sentence-transformers==2.7.0
pypdf==4.2.0
python-docx==1.1.2
pandas==2.2.2
openpyxl==3.1.2
numpy==1.26.4
pydantic==2.7.1
